# palabatu-be/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "internal_boulders.MergeObjection": {
                "properties": {
                    "body": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "merge_request_id": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_boulders.MergeRequest": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "resolved_at": {
                        "type": "string"
                    },
                    "resolved_by": {
                        "type": "string"
                    },
                    "source_boulder_id": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "suggested_by": {
                        "type": "string"
                    },
                    "target_boulder_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_boulders.MergeRequestListItem": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "objections": {
                        "items": {
                            "$ref": "#/components/schemas/internal_boulders.MergeObjection"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "reason": {
                        "type": "string"
                    },
                    "source_boulder_id": {
                        "type": "string"
                    },
                    "source_boulder_name": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "suggested_by": {
                        "type": "string"
                    },
                    "suggester_name": {
                        "type": "string"
                    },
                    "target_boulder_id": {
                        "type": "string"
                    },
                    "target_boulder_name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_boulders.ObjectToMergeRequest": {
                "properties": {
                    "body": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_boulders.ResolveMergeRequestRequest": {
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "override_hold": {
                        "type": "boolean"
                    },
                    "survivor_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "internal_boulders.SuggestMergeRequest": {
                "properties": {
                    "reason": {
                        "type": "string"
                    },
                    "target_boulder_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "palabatu-be_internal_apitypes.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "palabatu-be_internal_apitypes.SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and the JWT, e.g. \"Bearer eyJhbGc...\".",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Backend API for Palabatu, a community app for Indonesian bouldering enthusiasts (spot map, climber profiles, route/problem listings, comments, reactions, reports).",
        "title": "Palabatu API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/boulders/merge-requests": {
            "get": {
                "description": "Admin-only (Council/Associate title) review queue, with every objection filed against each request embedded.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/internal_boulders.MergeRequestListItem"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List pending boulder merge requests",
                "tags": [
                    "boulders"
                ]
            }
        },
        "/api/boulders/merge-requests/{id}/object": {
            "post": {
                "description": "Only the source or target boulder's own creator may object. Informs the admin's decision, does not block it.",
                "parameters": [
                    {
                        "description": "Merge request ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/internal_boulders.ObjectToMergeRequest",
                                        "summary": "body",
                                        "description": "Why this isn't the same rock"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Why this isn't the same rock",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_boulders.MergeObjection"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "request already resolved"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Object to a suggested merge (\"this is not the same rock\")",
                "tags": [
                    "boulders"
                ]
            }
        },
        "/api/boulders/merge-requests/{id}/resolve": {
            "post": {
                "description": "Admin-only. action=\"merge\" requires survivor_id (the admin's pick of which boulder survives) and is blocked by a 48h objection hold unless override_hold is set. action=\"reject\" needs no wait.",
                "parameters": [
                    {
                        "description": "Merge request ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/internal_boulders.ResolveMergeRequestRequest",
                                        "summary": "body",
                                        "description": "Decision"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Decision",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.SuccessResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "already resolved, or the 48h hold hasn't expired"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Resolve a boulder merge request",
                "tags": [
                    "boulders"
                ]
            }
        },
...
        "/api/boulders/{id}/merge-suggestions": {
            "post": {
                "description": "Any authenticated user may suggest a merge; executing one is admin-only (see resolve).",
                "parameters": [
                    {
                        "description": "Boulder ID (the suggested duplicate)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/internal_boulders.SuggestMergeRequest",
                                        "summary": "body",
                                        "description": "The other boulder, and why"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "The other boulder, and why",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/internal_boulders.MergeRequest"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/palabatu-be_internal_apitypes.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Suggest that two boulders are the same rock",
                "tags": [
                    "boulders"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:3001/"
        }
    ]
}